+++ modules/util/source-formatter/src/main/java/com/liferay/source/formatter/BaseSourceProcessor.java
@@ -37,7 +37,6 @@ import com.liferay.source.formatter.util.FileUtil;
 
 import java.io.File;
 import java.io.FileInputStream;
-import java.io.FileNotFoundException;
 import java.io.IOException;
 import java.io.InputStream;
 
@@ -49,13 +48,13 @@ import java.nio.charset.CodingErrorAction;
 
 import java.util.ArrayList;
 import java.util.Comparator;
-import java.util.Enumeration;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Objects;
 import java.util.Properties;
 import java.util.Set;
+import java.util.TreeSet;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.CopyOnWriteArrayList;
@@ -153,11 +152,11 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 	}
 
 	@Override
-	public List<SourceFormatterMessage> getSourceFormatterMessages() {
-		List<SourceFormatterMessage> sourceFormatterMessages =
-			new ArrayList<>();
+	public Set<SourceFormatterMessage> getSourceFormatterMessages() {
+		Set<SourceFormatterMessage> sourceFormatterMessages =
+			new TreeSet<>();
 
-		for (Map.Entry<String, List<SourceFormatterMessage>> entry :
+		for (Map.Entry<String, Set<SourceFormatterMessage>> entry :
 				_sourceFormatterMessagesMap.entrySet()) {
 
 			sourceFormatterMessages.addAll(entry.getValue());
@@ -173,17 +172,22 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 
 	@Override
 	public void processMessage(String fileName, String message, int lineCount) {
-		List<SourceFormatterMessage> sourceFormatterMessages =
+		Set<SourceFormatterMessage> sourceFormatterMessages =
 			_sourceFormatterMessagesMap.get(fileName);
 
 		if (sourceFormatterMessages == null) {
-			sourceFormatterMessages = new ArrayList<>();
+			sourceFormatterMessages = new TreeSet<>();
 		}
 
 		sourceFormatterMessages.add(
 			new SourceFormatterMessage(fileName, message, lineCount));
 
 		_sourceFormatterMessagesMap.put(fileName, sourceFormatterMessages);
+	}
+
+	@Override
+	public void setProperties(Properties properties) {
+		_properties = properties;
 	}
 
 	@Override
@@ -345,7 +349,9 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 
 			if (Objects.equals(value, defaultValue)) {
 				processMessage(
-					fileName, "No need to pass default value",
+					fileName,
+					"No need to pass default value '" + parametersList.get(1) +
+						"'",
 					getLineCount(content, matcher.start()));
 			}
 		}
@@ -371,79 +377,78 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		if (hasRedundantParentheses(ifClause, "||", "&&") ||
 			hasRedundantParentheses(ifClause, "&&", "||")) {
 
-			processMessage(fileName, "redundant parentheses", lineCount);
+			processMessage(
+				fileName, "Redundant parentheses in if-statement", lineCount);
 
 			return;
 		}
 
-		ifClause = stripRedundantParentheses(ifClause);
-
-		int level = 0;
-		int max = StringUtil.count(ifClause, CharPool.OPEN_PARENTHESIS);
 		int previousParenthesisPos = -1;
-
-		int[] levels = new int[max];
 
 		for (int i = 0; i < ifClause.length(); i++) {
 			char c = ifClause.charAt(i);
 
-			if ((c == CharPool.OPEN_PARENTHESIS) ||
-				(c == CharPool.CLOSE_PARENTHESIS)) {
+			if ((c != CharPool.OPEN_PARENTHESIS) &&
+				(c != CharPool.CLOSE_PARENTHESIS)) {
 
-				if (previousParenthesisPos != -1) {
-					String s = ifClause.substring(
-						previousParenthesisPos + 1, i);
+				continue;
+			}
 
-					if (hasMissingParentheses(s)) {
-						processMessage(
-							fileName, "missing parentheses", lineCount);
+			if (previousParenthesisPos != -1) {
+				String s = ifClause.substring(previousParenthesisPos + 1, i);
 
-						return;
-					}
+				if (hasMissingParentheses(s)) {
+					processMessage(
+						fileName, "Missing parentheses in if-statement",
+						lineCount);
+
+					return;
 				}
+			}
 
+			if (previousParenthesisPos == -1) {
 				previousParenthesisPos = i;
 
-				if (c == CharPool.OPEN_PARENTHESIS) {
-					levels[level] = i;
+				continue;
+			}
 
-					level += 1;
-				}
-				else {
-					int posOpenParenthesis = levels[level - 1];
+			previousParenthesisPos = i;
 
-					if (level > 1) {
-						char nextChar = ifClause.charAt(i + 1);
-						char previousChar = ifClause.charAt(
-							posOpenParenthesis - 1);
+			if (c != CharPool.OPEN_PARENTHESIS) {
+				continue;
+			}
 
-						if (!Character.isLetterOrDigit(nextChar) &&
-							(nextChar != CharPool.PERIOD) &&
-							!Character.isLetterOrDigit(previousChar)) {
+			char previousChar = ifClause.charAt(i - 1);
 
-							String s = ifClause.substring(
-								posOpenParenthesis + 1, i);
+			if ((previousChar != CharPool.OPEN_PARENTHESIS) &&
+				(previousChar != CharPool.SPACE)) {
 
-							if (hasRedundantParentheses(s)) {
-								processMessage(
-									fileName, "redundant parentheses",
-									lineCount);
+				continue;
+			}
 
-								return;
-							}
-						}
+			int j = i;
 
-						if ((previousChar == CharPool.OPEN_PARENTHESIS) &&
-							(nextChar == CharPool.CLOSE_PARENTHESIS)) {
+			while (true) {
+				j = ifClause.indexOf(StringPool.CLOSE_PARENTHESIS, j + 1);
 
-							processMessage(
-								fileName, "redundant parentheses", lineCount);
+				String s = ifClause.substring(i + 1, j);
 
-							return;
-						}
+				if (getLevel(s) == 0) {
+					char nextChar = ifClause.charAt(j + 1);
+
+					if (((previousChar == CharPool.OPEN_PARENTHESIS) &&
+						 (nextChar == CharPool.CLOSE_PARENTHESIS)) ||
+						(((nextChar == CharPool.CLOSE_PARENTHESIS) ||
+						  (nextChar == CharPool.SPACE)) &&
+						 hasRedundantParentheses(s))) {
+
+						processMessage(
+							fileName,
+							"Redundant parentheses in if-statement",
+							lineCount);
 					}
 
-					level -= 1;
+					break;
 				}
 			}
 		}
@@ -563,11 +568,11 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 				Properties bndFileLanguageProperties =
 					getBNDFileLanguageProperties(fileName);
 
-				if ((bndFileLanguageProperties == null) ||
+				if ((bndFileLanguageProperties != null) &&
 					!bndFileLanguageProperties.containsKey(languageKey)) {
 
 					processMessage(
-						fileName, "missing language key: " + languageKey);
+						fileName, "Missing language key '" + languageKey + "'");
 				}
 			}
 		}
@@ -591,9 +596,9 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 
 				StringBundler sb = new StringBundler(7);
 
-				sb.append("order ");
+				sb.append("Incorrect order '");
 				sb.append(elementName);
-				sb.append(StringPool.COLON);
+				sb.append("':");
 
 				if (Validator.isNotNull(parentElementName)) {
 					sb.append(StringPool.SPACE);
@@ -635,7 +640,8 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		if (content.contains("org.apache.commons.beanutils.PropertyUtils")) {
 			processMessage(
 				fileName,
-				"Do not use org.apache.commons.beanutils.PropertyUtils");
+				"Do not use org.apache.commons.beanutils.PropertyUtils, see " +
+					"LPS-62786");
 		}
 	}
 
@@ -650,7 +656,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			processMessage(
 				fileName,
 				"Use ResourceBundleUtil.getBundle instead of " +
-					"ResourceBundle.getBundle",
+					"ResourceBundle.getBundle, see LPS-58529",
 				lineCount);
 		}
 
@@ -658,7 +664,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			processMessage(
 				fileName,
 				"Use ResourceBundleUtil.getString instead of " +
-					"resourceBundle.getString",
+					"resourceBundle.getString, see LPS-58529",
 				lineCount);
 		}
 	}
@@ -796,7 +802,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 
 			content = StringUtil.replace(content, _oldCopyright, copyright);
 
-			processMessage(fileName, "old (c)");
+			processMessage(fileName, "File contains old copyright information");
 		}
 
 		if (!content.contains(copyright)) {
@@ -1098,7 +1104,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		}
 
 		fileName = StringUtil.replace(
-			fileName, StringPool.BACK_SLASH, StringPool.SLASH);
+			fileName, CharPool.BACK_SLASH, CharPool.SLASH);
 
 		File file = new File(fileName);
 
@@ -1172,7 +1178,9 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 				}
 
 				if (delimeter != CharPool.AMPERSAND) {
-					processMessage(fileName, "delimeter", lineCount);
+					processMessage(
+						fileName, "Incorrect delimeter '" + delimeter + "'",
+						lineCount);
 				}
 
 				return line;
@@ -1321,14 +1329,14 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			return line;
 		}
 
-		String whitespace = matcher.group(1);
+		String whitespace = matcher.group(2);
 
 		if (whitespace.length() > 0) {
 			return line;
 		}
 
-		return line.substring(0, matcher.start(1)) + StringPool.SPACE +
-			line.substring(matcher.start(1));
+		return line.substring(0, matcher.start(2)) + StringPool.SPACE +
+			line.substring(matcher.start(2));
 	}
 
 	protected String formatIncorrectSyntax(
@@ -1375,6 +1383,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		else {
 			javaSourceProcessor = new JavaSourceProcessor();
 
+			javaSourceProcessor.setProperties(_properties);
 			javaSourceProcessor.setSourceFormatterArgs(sourceFormatterArgs);
 		}
 
@@ -1422,7 +1431,8 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			}
 
 			processMessage(
-				fileName, "plus", getLineCount(content, matcher.start(1)));
+				fileName, "Incorrect use of '+' inside StringBundler",
+				getLineCount(content, matcher.start(1)));
 		}
 
 		matcher = sbAppendWithStartingSpacePattern.matcher(content);
@@ -1438,7 +1448,8 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 				(getLineLength(firstLine) >= maxLineLength)) {
 
 				processMessage(
-					fileName, "leading space in sb",
+					fileName,
+					"Do not append string starting with space to StringBundler",
 					getLineCount(content, matcher.start(3)));
 			}
 			else {
@@ -1510,10 +1521,10 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			linePart = formatIncorrectSyntax(linePart, " }", "}", false);
 			linePart = formatIncorrectSyntax(linePart, " )", ")", false);
 			linePart = formatIncorrectSyntax(linePart, "( ", "(", false);
-			linePart = formatIncorrectSyntax(linePart, "\\w( ?)=");
-			linePart = formatIncorrectSyntax(linePart, "=( ?)\\w");
-			linePart = formatIncorrectSyntax(linePart, "for \\(.*( ?):");
-			linePart = formatIncorrectSyntax(linePart, "for \\(.*:( ?).+");
+			linePart = formatIncorrectSyntax(linePart, "\\w(( ?)=)");
+			linePart = formatIncorrectSyntax(linePart, "(=( ?))\\w");
+			linePart = formatIncorrectSyntax(linePart, "for \\(.*(( ?):)");
+			linePart = formatIncorrectSyntax(linePart, "for \\(.*(:( ?)).+");
 		}
 
 		if (!linePart.startsWith("##")) {
@@ -1662,7 +1673,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			getBNDFileLocationAndContentTuple(fileName);
 
 		if (bndFileLocationAndContentTuple == null) {
-			return null;
+			return new Properties();
 		}
 
 		String bndFileLocation =
@@ -1676,6 +1687,17 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 
 		String bndContent = (String)bndFileLocationAndContentTuple.getObject(1);
 
+		if (bndContent.matches(
+				"[\\s\\S]*Provide-Capability:.*liferay\\.resource\\.bundle" +
+					"[\\s\\S]*")) {
+
+			// Return null, in order to skip checking for language keys for
+			// modules that use LanguageExtender. No fix in place for this right
+			// now.
+
+			return null;
+		}
+
 		Matcher matcher = bndContentDirPattern.matcher(bndContent);
 
 		if (matcher.find()) {
@@ -1683,7 +1705,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 				bndFileLocation + matcher.group(1) + "/Language.properties");
 
 			if (!file.exists()) {
-				return null;
+				return new Properties();
 			}
 
 			properties = new Properties();
@@ -1697,7 +1719,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			return properties;
 		}
 
-		return null;
+		return new Properties();
 	}
 
 	protected Tuple getBNDFileLocationAndContentTuple(String fileName)
@@ -1773,10 +1795,10 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			String content = FileUtil.read(file);
 
 			fileName = StringUtil.replace(
-				fileName, StringPool.BACK_SLASH, StringPool.SLASH);
+				fileName, CharPool.BACK_SLASH, CharPool.SLASH);
 
 			fileName = StringUtil.replace(
-				fileName, StringPool.SLASH, StringPool.PERIOD);
+				fileName, CharPool.SLASH, CharPool.PERIOD);
 
 			int pos = fileName.indexOf("com.");
 
@@ -1836,18 +1858,8 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 	}
 
 	protected File getFile(String fileName, int level) {
-		for (int i = 0; i < level; i++) {
-			File file = new File(
-				sourceFormatterArgs.getBaseDirName() + fileName);
-
-			if (file.exists()) {
-				return file;
-			}
-
-			fileName = "../" + fileName;
-		}
-
-		return null;
+		return _sourceFormatterHelper.getFile(
+			sourceFormatterArgs.getBaseDirName(), fileName, level);
 	}
 
 	protected List<String> getFileNames(
@@ -1855,12 +1867,24 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 			String[] excludes, String[] includes)
 		throws Exception {
 
+		return getFileNames(
+			basedir, recentChangesFileNames, excludes, includes,
+			sourceFormatterArgs.isIncludeSubrepositories());
+	}
+
+	protected List<String> getFileNames(
+			String basedir, List<String> recentChangesFileNames,
+			String[] excludes, String[] includes,
+			boolean includeSubrepositories)
+		throws Exception {
+
 		if (_excludes != null) {
 			excludes = ArrayUtil.append(excludes, _excludes);
 		}
 
 		return _sourceFormatterHelper.getFileNames(
-			basedir, recentChangesFileNames, excludes, includes);
+			basedir, recentChangesFileNames, excludes, includes,
+			includeSubrepositories);
 	}
 
 	protected List<String> getFileNames(
@@ -2079,21 +2103,59 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		return x + 1;
 	}
 
-	protected ComparableVersion getMainReleaseComparableVersion() {
-		if (_mainReleaseComparableVersion != null) {
-			return _mainReleaseComparableVersion;
+	protected ComparableVersion getMainReleaseComparableVersion(
+			String fileName, String absolutePath, boolean checkModuleVersion)
+		throws Exception {
+
+		boolean usePortalReleaseVersion = true;
+
+		if (checkModuleVersion &&
+			(!portalSource || isModulesFile(absolutePath))) {
+
+			usePortalReleaseVersion = false;
 		}
 
-		String releaseVersion = ReleaseInfo.getVersion();
+		String releaseVersion = StringPool.BLANK;
+
+		if (usePortalReleaseVersion) {
+			if (_mainReleaseComparableVersion != null) {
+				return _mainReleaseComparableVersion;
+			}
+
+			releaseVersion = ReleaseInfo.getVersion();
+		}
+		else {
+			Tuple bndFileLocationAndContentTuple =
+				getBNDFileLocationAndContentTuple(fileName);
+
+			if (bndFileLocationAndContentTuple == null) {
+				return null;
+			}
+
+			String bndContent =
+				(String)bndFileLocationAndContentTuple.getObject(1);
+
+			Matcher matcher = bndReleaseVersionPattern.matcher(bndContent);
+
+			if (!matcher.find()) {
+				return null;
+			}
+
+			releaseVersion = matcher.group(1);
+		}
 
 		int pos = releaseVersion.lastIndexOf(CharPool.PERIOD);
 
 		String mainReleaseVersion = releaseVersion.substring(0, pos) + ".0";
 
-		_mainReleaseComparableVersion = new ComparableVersion(
+		ComparableVersion mainReleaseComparableVersion = new ComparableVersion(
 			mainReleaseVersion);
 
-		return _mainReleaseComparableVersion;
+		if (usePortalReleaseVersion) {
+			_mainReleaseComparableVersion = mainReleaseComparableVersion;
+		}
+
+		return mainReleaseComparableVersion;
 	}
 
 	protected List<String> getModuleLangDirNames(
@@ -2589,7 +2651,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		}
 
 		if (sourceFormatterArgs.isPrintErrors()) {
-			List<SourceFormatterMessage> sourceFormatterMessages =
+			Set<SourceFormatterMessage> sourceFormatterMessages =
 				_sourceFormatterMessagesMap.get(fileName);
 
 			if (sourceFormatterMessages != null) {
@@ -2799,28 +2861,6 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		return sb.toString();
 	}
 
-	protected String stripRedundantParentheses(String s) {
-		for (int x = 0;;) {
-			x = s.indexOf(CharPool.OPEN_PARENTHESIS, x + 1);
-			int y = s.indexOf(CharPool.CLOSE_PARENTHESIS, x);
-
-			if ((x == -1) || (y == -1)) {
-				return s;
-			}
-
-			String linePart = s.substring(x + 1, y);
-
-			linePart = StringUtil.replace(
-				linePart, StringPool.COMMA, StringPool.BLANK);
-
-			if (Validator.isAlphanumericName(linePart) ||
-				Validator.isNull(linePart)) {
-
-				s = s.substring(0, x) + s.substring(y + 1);
-			}
-		}
-	}
-
 	protected String trimContent(String content, boolean allowLeadingSpaces)
 		throws IOException {
 
@@ -2890,6 +2930,8 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		"[a-z]+[-_a-zA-Z0-9]*");
 	protected static Pattern bndContentDirPattern = Pattern.compile(
 		"\\scontent=(.*?)(,\\\\|\n|$)");
+	protected static Pattern bndReleaseVersionPattern = Pattern.compile(
+		"Bundle-Version: (.*)\n");
 	protected static Pattern emptyCollectionPattern = Pattern.compile(
 		"Collections\\.EMPTY_(LIST|MAP|SET)");
 	protected static Pattern getterUtilGetPattern = Pattern.compile(
@@ -2949,93 +2991,16 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 		return excludesList.toArray(new String[excludesList.size()]);
 	}
 
-	private Properties _getProperties() throws Exception {
-		String fileName = "source-formatter.properties";
-
-		Properties properties = new Properties();
-
-		List<Properties> propertiesList = new ArrayList<>();
-
-		int level = 2;
-
-		if (portalSource) {
-			level = PORTAL_MAX_DIR_LEVEL;
-		}
-
-		for (int i = 0; i <= level; i++) {
-			try {
-				InputStream inputStream = new FileInputStream(
-					sourceFormatterArgs.getBaseDirName() + fileName);
-
-				Properties props = new Properties();
-
-				props.load(inputStream);
-
-				propertiesList.add(props);
-
-				break;
-			}
-			catch (FileNotFoundException fnfe) {
-			}
-
-			fileName = "../" + fileName;
-		}
-
-		if (propertiesList.isEmpty()) {
-			return properties;
-		}
-
-		properties = propertiesList.get(0);
-
-		if (propertiesList.size() == 1) {
-			return properties;
-		}
-
-		for (int i = 1; i < propertiesList.size(); i++) {
-			Properties props = propertiesList.get(i);
-
-			Enumeration<String> enu =
-				(Enumeration<String>)props.propertyNames();
-
-			while (enu.hasMoreElements()) {
-				String key = enu.nextElement();
-
-				String value = props.getProperty(key);
-
-				if (Validator.isNull(value)) {
-					continue;
-				}
-
-				if (key.contains("excludes")) {
-					String existingValue = properties.getProperty(key);
-
-					if (Validator.isNotNull(existingValue)) {
-						value = existingValue + StringPool.COMMA + value;
-					}
-
-					properties.put(key, value);
-				}
-				else if (!properties.containsKey(key)) {
-					properties.put(key, value);
-				}
-			}
-		}
-
-		return properties;
-	}
-
 	private void _init() {
-		portalSource = _isPortalSource();
-
-		_sourceFormatterMessagesMap = new HashMap<>();
-
 		try {
-			_properties = _getProperties();
-
 			_sourceFormatterHelper = new SourceFormatterHelper(
 				sourceFormatterArgs.isUseProperties());
 
 			_sourceFormatterHelper.init();
+
+			portalSource = _isPortalSource();
+
+			_sourceFormatterMessagesMap = new HashMap<>();
 		}
 		catch (Exception e) {
 			ReflectionUtil.throwException(e);
@@ -3110,7 +3075,7 @@ public abstract class BaseSourceProcessor implements SourceProcessor {
 	private Properties _properties;
 	private List<String> _runOutsidePortalExcludes;
 	private SourceFormatterHelper _sourceFormatterHelper;
-	private Map<String, List<SourceFormatterMessage>>
+	private Map<String, Set<SourceFormatterMessage>>
 		_sourceFormatterMessagesMap = new ConcurrentHashMap<>();
 	private boolean _usePortalCompatImport;