+++ portal-kernel/src/com/liferay/portal/kernel/util/StringUtil.java
@@ -1124,7 +1124,7 @@ public class StringUtil {
 					continue;
 				}
 
-				if ((i + texts[j].length() <= toIndex + 1) &&
+				if (((i + texts[j].length()) <= (toIndex + 1)) &&
 					s.startsWith(texts[j], i)) {
 
 					return i;
@@ -1556,7 +1556,7 @@ public class StringUtil {
 					continue;
 				}
 
-				if ((i + texts[j].length() <= toIndex + 1) &&
+				if (((i + texts[j].length()) <= (toIndex + 1)) &&
 					s.startsWith(texts[j], i)) {
 
 					return i;
@@ -2303,6 +2303,7 @@ public class StringUtil {
 				sb.append(s.substring(x, y));
 
 				x = y + 1;
+
 				y = s.indexOf(oldSub, x);
 			}
 
@@ -2471,6 +2472,7 @@ public class StringUtil {
 				sb.append(s.substring(x, y));
 
 				x = y + length;
+
 				y = s.indexOf(oldSub, x);
 			}
 
@@ -2713,6 +2715,7 @@ public class StringUtil {
 				sb.append(newSub);
 
 				x = y + length;
+
 				y = s.indexOf(oldSub, x);
 			}
 
@@ -3152,6 +3155,7 @@ public class StringUtil {
 
 		while (true) {
 			int x = s.indexOf(begin, pos);
+
 			int y = s.indexOf(end, x + begin.length());
 
 			if ((x == -1) || (y == -1)) {
@@ -3218,6 +3222,7 @@ public class StringUtil {
 
 		while (true) {
 			int x = s.indexOf(begin, pos);
+
 			int y = s.indexOf(end, x + begin.length());
 
 			if ((x == -1) || (y == -1)) {
@@ -3259,6 +3264,7 @@ public class StringUtil {
 		}
 
 		char[] chars = s.toCharArray();
+
 		char[] reverse = new char[chars.length];
 
 		for (int i = 0; i < chars.length; i++) {
@@ -3399,19 +3405,21 @@ public class StringUtil {
 			return null;
 		}
 
-		if (s.length() <= length) {
+		if (s.codePointCount(0, s.length()) <= length) {
 			return s;
 		}
 
 		if (length < suffix.length()) {
-			return s.substring(0, length);
+			return s.substring(0, s.offsetByCodePoints(0, length));
 		}
 
 		int curLength = length;
 
-		for (int j = (curLength - suffix.length()); j >= 0; j--) {
-			if (Character.isWhitespace(s.charAt(j))) {
-				curLength = j;
+		for (int j = (curLength - suffix.length() + 1), offset; j > 0; j--) {
+			offset = s.offsetByCodePoints(0, j);
+
+			if (Character.isWhitespace(s.codePointBefore(offset))) {
+				curLength = j - 1;
 
 				break;
 			}
@@ -3421,7 +3429,7 @@ public class StringUtil {
 			curLength = length - suffix.length();
 		}
 
-		String temp = s.substring(0, curLength);
+		String temp = s.substring(0, s.offsetByCodePoints(0, curLength));
 
 		return temp.concat(suffix);
 	}
@@ -3543,12 +3551,14 @@ public class StringUtil {
 		List<String> nodeValues = new ArrayList<>();
 
 		int offset = 0;
+
 		int pos = s.indexOf(delimiter, offset);
 
 		while (pos != -1) {
 			nodeValues.add(s.substring(offset, pos));
 
 			offset = pos + 1;
+
 			pos = s.indexOf(delimiter, offset);
 		}
 
@@ -3677,12 +3687,14 @@ public class StringUtil {
 		List<String> nodeValues = new ArrayList<>();
 
 		int offset = 0;
+
 		int pos = s.indexOf(delimiter, offset);
 
 		while (pos != -1) {
 			nodeValues.add(s.substring(offset, pos));
 
 			offset = pos + delimiter.length();
+
 			pos = s.indexOf(delimiter, offset);
 		}
 
@@ -3707,6 +3719,7 @@ public class StringUtil {
 	 */
 	public static boolean[] split(String s, String delimiter, boolean x) {
 		String[] array = split(s, delimiter);
+
 		boolean[] newArray = new boolean[array.length];
 
 		for (int i = 0; i < array.length; i++) {
@@ -3739,6 +3752,7 @@ public class StringUtil {
 	 */
 	public static double[] split(String s, String delimiter, double x) {
 		String[] array = split(s, delimiter);
+
 		double[] newArray = new double[array.length];
 
 		for (int i = 0; i < array.length; i++) {
@@ -3770,6 +3784,7 @@ public class StringUtil {
 	 */
 	public static float[] split(String s, String delimiter, float x) {
 		String[] array = split(s, delimiter);
+
 		float[] newArray = new float[array.length];
 
 		for (int i = 0; i < array.length; i++) {
@@ -3801,6 +3816,7 @@ public class StringUtil {
 	 */
 	public static int[] split(String s, String delimiter, int x) {
 		String[] array = split(s, delimiter);
+
 		int[] newArray = new int[array.length];
 
 		for (int i = 0; i < array.length; i++) {
@@ -3832,6 +3848,7 @@ public class StringUtil {
 	 */
 	public static long[] split(String s, String delimiter, long x) {
 		String[] array = split(s, delimiter);
+
 		long[] newArray = new long[array.length];
 
 		for (int i = 0; i < array.length; i++) {
@@ -3863,6 +3880,7 @@ public class StringUtil {
 	 */
 	public static short[] split(String s, String delimiter, short x) {
 		String[] array = split(s, delimiter);
+
 		short[] newArray = new short[array.length];
 
 		for (int i = 0; i < array.length; i++) {
@@ -4149,6 +4167,7 @@ public class StringUtil {
 
 		while (true) {
 			int x = s.indexOf(begin, pos);
+
 			int y = s.indexOf(end, x + begin.length());
 
 			if ((x == -1) || (y == -1)) {
@@ -4593,6 +4612,7 @@ public class StringUtil {
 		}
 
 		int len = s.length();
+
 		int x = len;
 
 		for (int i = 0; i < len; i++) {
@@ -4646,6 +4666,7 @@ public class StringUtil {
 		}
 
 		int len = s.length();
+
 		int x = len;
 
 		for (int i = 0; i < len; i++) {
@@ -4707,6 +4728,7 @@ public class StringUtil {
 		}
 
 		int len = s.length();
+
 		int x = len;
 
 		for (int i = 0; i < len; i++) {