+++ portal-kernel/test/unit/com/liferay/portal/kernel/servlet/ServletResponseUtilRangeTest.java
@@ -226,6 +226,7 @@ public class ServletResponseUtilRangeTest extends PowerMockito {
 					throws Throwable {
 
 					Object[] args = invocation.getArguments();
+
 					File file = (File)args[0];
 
 					return file.delete();
@@ -308,6 +309,7 @@ public class ServletResponseUtilRangeTest extends PowerMockito {
 			byte[] bytes = ArrayUtil.subset(content, start, end + 1);
 
 			Assert.assertArrayEquals(bytes, lines[3].getBytes("UTF-8"));
+
 			Assert.assertEquals(StringPool.DOUBLE_DASH, lines[4]);
 		}
 	}