+++ modules/sdk/gradle-plugins-defaults/CHANGELOG.markdown
@@ -111,27 +111,210 @@ an OSGi project that does not contain a `packageinfo` file.
 ## 1.2.3 - 2016-09-01
 
 ### Changed
-- [LPS-67863]: The file which contains the version overrides for an OSGi module
+- [LPS-67863]: The file that contains the version overrides for an OSGi module
 is now called `.version-override-${project.name}.properties`.
-- [LPS-67863]: The `packageinfo` versions are always overriden with the versions
-specified in the `.version-override-${project.name}.properties` file, even if
-the ones in the `packageinfo` files are greater.
+- [LPS-67863]: The `packageinfo` versions are always overridden with the
+versions specified in the `.version-override-${project.name}.properties` file,
+even if the versions in the `packageinfo` files are greater.
 
 ## 1.2.4 - 2016-09-01
 
 ### Changed
-- [LPS-67863]: In order to prevent releasing modules with unpublished
-dependencies, Disable the `printArtifactPublishCommands` task if the project's
-`build.gradle` contains the string `version: "default"`.
-- [LPS-67863]: The `.version-override-${project.name}.properties` now contains
-only the version overrides that are different from the versions specified in the
-`bnd.bnd` and `packageinfo` files.
+- [LPS-67863]: Disable the `printArtifactPublishCommands` task if the project's
+`build.gradle` contains the string `version: "default"`, to prevent releasing
+modules with unpublished dependencies.
+- [LPS-67863]: The `.version-override-${project.name}.properties` file now
+contains only the version overrides that differ from the versions specified in
+the `bnd.bnd` and `packageinfo` files.
+
+## 1.2.5 - 2016-09-01
+
+### Fixed
+- [LPS-67863]: Avoid throwing an exception while running `gradlew baseline
+-PsyncRelease` on a project that does not contain a
+`.version-override-${project.name}.properties` file.
+
+## 1.2.6 - 2016-09-02
+
+### Fixed
+- [LPS-67863]: Avoid throwing an exception while running the
+`printArtifactPublishCommands` task on a project that does not contain a
+`build.gradle` file.
+
+## 1.2.7 - 2016-09-02
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.21.
+
+## 1.2.8 - 2016-09-02
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.22.
+
+## 1.2.9 - 2016-09-05
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.23.
+
+## 1.2.10 - 2016-09-05
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.24.
+
+## 1.2.11 - 2016-09-06
+
+### Changed
+- [LPS-67996]: Update the [Liferay Gradle Plugins] dependency to version 2.0.25.
+
+## 1.2.12 - 2016-09-07
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.26.
+
+## 1.2.13 - 2016-09-07
+
+### Fixed
+- [LPS-68009]: Reject snapshot artifacts while resolving the `baseline`
+configuration.
+
+## 1.2.14 - 2016-09-07
+
+### Changed
+- [LPS-68014]: Update the [Liferay Gradle Plugins] dependency to version 2.0.27.
+
+## 1.2.15 - 2016-09-08
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.28.
+
+## 1.2.16 - 2016-09-08
+
+### Added
+- [LPS-67863]: Allow dependency versions to be overridden in the
+`.version-override-${project.name}.properties` file:
+
+		[artifact group]-[artifact name]=[new version]
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.29.
+
+## 1.2.17 - 2016-09-09
+
+### Changed
+- [LPS-61099]: Update the [Liferay Gradle Plugins] dependency to version 2.0.30.
+- [LRDOCS-2841]: Look for the `.releng` directory starting from the project
+directory instead of the root project directory. Doing this lets submodules like
+`content-targeting` have their own separate `.releng` directory.
+
+## 1.2.18 - 2016-09-12
+
+### Changed
+- [LPS-67766]: Update the [Liferay Gradle Plugins] dependency to version 2.0.31.
+
+## 1.2.19 - 2016-09-13
+
+### Changed
+- [LPS-67352]: Update the [Liferay Gradle Plugins] dependency to version 2.0.32.
+
+## 1.2.20 - 2016-09-13
+
+### Changed
+- [LPS-67986]: Update the [Liferay Gradle Plugins] dependency to version 2.0.33.
+
+## 1.2.21 - 2016-09-13
+
+### Changed
+- [LRDOCS-2981]: Prepend *Module* string to `appJavadoc` module headings.
+
+## 1.2.22 - 2016-09-14
+
+### Changed
+- [LPS-68131]: Update the [Liferay Gradle Plugins] dependency to version 2.0.34.
+
+## 1.2.23 - 2016-09-16
+
+### Changed
+- [LPS-68131]: Update the [Liferay Gradle Plugins] dependency to version 2.0.35.
+
+## 1.2.24 - 2016-09-20
+
+### Changed
+- [LPS-67352]: Update the [Liferay Gradle Plugins] dependency to version 2.0.36.
+- [LPS-68230]: Configure [Liferay Gradle Plugins Node] to use version 6.6.0 of
+Node.js.
+
+### Removed
+- [LPS-68230]: To reduce the number of plugins applied to a project and improve
+performance, plugins in `com.liferay.gradle.plugins.defaults.internal` are no
+longer applied via `apply plugin`.
+
+## 1.2.25 - 2016-09-20
+
+### Changed
+- [LPS-66906]: Update the [Liferay Gradle Plugins] dependency to version 2.0.37.
+
+## 1.2.26 - 2016-09-21
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.38.
+
+## 1.2.27 - 2016-09-22
+
+### Changed
+- [LPS-68305]: Set the `buildService.buildNumberIncrement` property to `false`
+by default.
+
+## 1.2.28 - 2016-09-22
+
+### Changed
+- [LPS-68297]: Update the [Liferay Gradle Plugins] dependency to version 2.0.39.
+
+## 1.2.29 - 2016-09-22
+
+### Added
+- [LPS-66906]: When on Jenkins, override the [`sass-binary-path`](https://github.com/sass/node-sass#binary-configuration-parameters)
+argument in the `npmInstall` task with the value of the project property
+`nodejs.npm.ci.sass.binary.site`.
+
+### Changed
+- [LPS-66906]: Update the [Liferay Gradle Plugins] dependency to version 2.0.40.
+
+## 1.2.30 - 2016-09-23
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.41.
+
+## 1.2.31 - 2016-09-23
+
+### Added
+- [LPS-68306]: Set the system property `portal.pre.build` to `true` to include
+only the projects that contain a `.lfrbuild-portal-pre` marker file.
+
+## 1.2.32 - 2016-09-26
+
+### Changed
+- [LPS-66853]: Update the [Liferay Gradle Plugins] dependency to version 2.0.42.
 
 [Liferay Gradle Plugins]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins
+[Liferay Gradle Plugins Node]: https://github.com/liferay/liferay-portal/tree/master/modules/sdk/gradle-plugins-node
+[LPS-61099]: https://issues.liferay.com/browse/LPS-61099
 [LPS-66853]: https://issues.liferay.com/browse/LPS-66853
+[LPS-66906]: https://issues.liferay.com/browse/LPS-66906
 [LPS-67023]: https://issues.liferay.com/browse/LPS-67023
 [LPS-67352]: https://issues.liferay.com/browse/LPS-67352
 [LPS-67658]: https://issues.liferay.com/browse/LPS-67658
 [LPS-67694]: https://issues.liferay.com/browse/LPS-67694
+[LPS-67766]: https://issues.liferay.com/browse/LPS-67766
 [LPS-67804]: https://issues.liferay.com/browse/LPS-67804
 [LPS-67863]: https://issues.liferay.com/browse/LPS-67863
+[LPS-67986]: https://issues.liferay.com/browse/LPS-67986
+[LPS-67996]: https://issues.liferay.com/browse/LPS-67996
+[LPS-68009]: https://issues.liferay.com/browse/LPS-68009
+[LPS-68014]: https://issues.liferay.com/browse/LPS-68014
+[LPS-68131]: https://issues.liferay.com/browse/LPS-68131
+[LPS-68230]: https://issues.liferay.com/browse/LPS-68230
+[LPS-68305]: https://issues.liferay.com/browse/LPS-68305
+[LPS-68297]: https://issues.liferay.com/browse/LPS-68297
+[LPS-68306]: https://issues.liferay.com/browse/LPS-68306
+[LRDOCS-2841]: https://issues.liferay.com/browse/LRDOCS-2841
+[LRDOCS-2981]: https://issues.liferay.com/browse/LRDOCS-2981